+++ tablet_qt/diagnosis/icd10.cpp
@@ -36,8 +36,10 @@ const QString Icd10::XSTRING_TASKNAME("icd10");
 // Main functions
 // ============================================================================
 
-Icd10::Icd10(CamcopsApp& app, QObject* parent) :
-    DiagnosticCodeSet(app, XSTRING_TASKNAME, tr("ICD-10"), parent)
+Icd10::Icd10(CamcopsApp& app, QObject* parent,
+             bool dummy_creation_no_xstrings) :
+    DiagnosticCodeSet(app, XSTRING_TASKNAME, tr("ICD-10"),
+                      parent, dummy_creation_no_xstrings)
 {
     m_creation_stack.push(DepthItemPair(0, nullptr));  // root: depth 0, no parent
     addIcd10Codes(BASE_CODES);