+++ tablet_qt/diagnosis/icd9cm.h
@@ -31,7 +31,8 @@ class Icd9cm : public DiagnosticCodeSet
     Q_OBJECT
 
 public:
-    Icd9cm(CamcopsApp& app);
+    Icd9cm(CamcopsApp& app, QObject* parent = nullptr,
+           bool dummy_creation_no_xstrings = false);
 
     using CodeDescriptionPair = QPair<QString, QString>;
     using DepthItemPair = QPair<int, DiagnosticCode*>;